glove_relay/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "glove-relay"
version = "3.0.0"
description = "EdgeAI DataGlove V3 Relay Server — bridges ESP32 protobuf UDP ↔ React WebSocket frontend with on-device L1/L2 inference, NLP grammar correction, and TTS."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "EdgeAI DataGlove Team"},
]
keywords = ["sign-language", "glove", "relay", "stgcn", "edge-ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "websockets>=11.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "protobuf>=4.24.0",
    "edge-tts>=6.1.0",
    "numpy>=1.24.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0.1",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
    "ruff>=0.1.0",
]

[project.scripts]
glove-relay = "src.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 120
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
